split up analysis functions vs plotting functions
